rename ncdr_reference -> ncdr
