+++ deployment/roles/ncdr/tasks/main.yml
@@ -51,7 +51,7 @@
 - name: Create local settings
   template:
     src: "templates/local_settings.py.jinja2"
-    dest: "{{ PROJECT_PATH }}/ncdr_reference/local_settings.py"
+    dest: "{{ PROJECT_PATH }}/ncdr/local_settings.py"
     mode: 0744
     owner: ubuntu
