+++ ncdr/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for ncdr_reference project.
+Django settings for ncdr project.
 
 Generated by 'django-admin startproject' using Django 1.11.6.
 
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'markdown_deux',
     'bootstrapform',
     'rest_framework',
-    'ncdr_reference',
+    'ncdr',
 ]
 
 MIDDLEWARE = [
@@ -57,7 +57,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'ncdr_reference.urls'
+ROOT_URLCONF = 'ncdr.urls'
 
 STATIC_ROOT = 'static'
 
@@ -88,14 +88,14 @@ TEMPLATES = [
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
-                'ncdr_reference.context_processors.settings',
-                'ncdr_reference.context_processors.models',
+                'ncdr.context_processors.settings',
+                'ncdr.context_processors.models',
             ],
         },
     },
 ]
 
-WSGI_APPLICATION = 'ncdr_reference.wsgi.application'
+WSGI_APPLICATION = 'ncdr.wsgi.application'
 
 
 # Database
@@ -149,6 +149,6 @@ DATA_UPLOAD_MAX_NUMBER_FIELDS = 100000
 STATIC_URL = '/static/'
 if 'test' not in sys.argv:
     try:
-        from ncdr_reference.local_settings import *  # noqa: F401,F403
+        from ncdr.local_settings import *  # noqa: F401,F403
     except ImportError:
         pass
