+++ ncdr/urls.py
@@ -1,4 +1,4 @@
-"""ncdr_reference URL Configuration
+"""ncdr URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/1.11/topics/http/urls/
